requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argcomplete==1.10.3
asttokens==3.0.1
attrs==26.1.0
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.8.2
Bottleneck==1.6.0
build==1.4.2
cachetools==4.2.4
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.1.8
coloredlogs==15.0.1
comm==0.2.3
compressed-rtf==1.0.7
cryptography==46.0.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
docx2txt==0.9
durationpy==0.10
ebcdic==2.0.1
exceptiongroup==1.3.1
executing==2.2.1
extract-msg==0.28.7
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fsspec==2026.3.0
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-aiplatform==1.146.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-texttospeech==2.36.0
google-cloud-translate==2.0.1
google-cloud-vectorsearch==0.9.0
google-cloud-vision==3.13.0
google-crc32c==1.8.0
google-genai==1.72.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.0
googletrans==4.0.2
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.48.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
IMAPClient==2.1.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-chroma==1.1.0
langchain-core==1.2.27
langchain-google-genai==4.2.1
langchain-google-vertexai==3.2.2
langchain-tests==1.1.5
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.26
lxml==6.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
moviepy==2.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.2
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20191110
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.9.4
pluggy==1.6.0
proglog==0.1.12
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.3.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-pptx==0.6.23
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.4.0
shellingham==1.5.4
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SpeechRecognition==3.8.1
stack-data==0.6.3
sympy==1.14.0
syrupy==5.1.0
tenacity==9.1.4
textract==1.6.5
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
validators==0.35.0
vcrpy==8.1.1
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
xlrd==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
